requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==3.2.9
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-session-timeout==0.1.0
idna==3.4
Pillow==9.3.0
pyngrok==5.2.1
python-decouple==3.6
pytz==2022.7
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
